Rename lib header to appimage_info.h
