+++ include/appimage/appimage_info.h
