+++ src/lib/appimage_info.cpp
@@ -3,7 +3,7 @@
 //
 
 
-#include "appimage/appimageinfo.h"
+#include "appimage/appimage_info.h"
 #include "../entities/FileMetadataExtractor.h"
 
 void extract_desktop_file(const char *appimage_path, const char *output_path) {
